added working login page and sign up page
