+++ src/SignUp.jsx
@@ -1,27 +1,23 @@
 import React, {useState} from 'react';
 import {auth} from './firebase-config';
-import {signInWithEmailAndPassword} from '@firebase/auth';
-import logo from './UMKC_Logo.png';
+import {createUserWithEmailAndPassword} from '@firebase/auth';
 
-function Login(){
+function SignUp(){
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
 
-    const signIn = (event) => {
+    const signUp = (event) => {
         event.preventDefault();
-        signInWithEmailAndPassword(auth, email, password).then((userCredential) => {
+        createUserWithEmailAndPassword(auth, email, password).then((userCredential) => {
             console.log(userCredential);
         }).catch((error) => {
             console.log(error);
         })
     }
     return(
-        <div className='sign-in-container'>
-            <div className="logo">
-                <img src={logo} alt="UMKC" />
-            </div>
-            <form onSubmit={signIn}>
-                <h1>Log In</h1>
+        <div className='sign-up-container'>
+            <form onSubmit={signUp}>
+                <h1>Create Account</h1>
                 <input 
                 type="email" 
                 placeholder='Enter your email...' 
@@ -32,9 +28,9 @@ function Login(){
                 placeholder='Enter your password...' 
                 value={password}
                 onChange={(event) => {setPassword(event.target.value)}}/>
-                <button type="submit">Log In</button>
+                <button type="submit">Sign Up</button>
             </form>
         </div>
     )
 }
-export default Login;
+export default SignUp;
