+++ src/firebase-config.js
@@ -1,6 +1,7 @@
 import { initializeApp } from "firebase/app";
 import {getFirestore} from '@firebase/firestore';
 import {getStorage} from '@firebase/storage';
+import {getAuth} from '@firebase/auth';
 
 const firebaseConfig = {
   apiKey: "AIzaSyCh2X2opN14ARAoka-hMeVvPzrYoQ_UCwk",
@@ -16,3 +17,4 @@ const app = initializeApp(firebaseConfig);
 
 export const db = getFirestore(app);
 export const storage = getStorage(app);
+export const auth = getAuth(app);
